added FA2Helper namespace
