+++ src/index.ts
@@ -26,7 +26,6 @@ export * from './chain/tezos/contracts/tzip12/ChainlinkTokenHelper';
 export * from './chain/tezos/contracts/tzip12/MultiAssetTokenHelper';
 export * from './chain/tezos/contracts/tzip12/SingleAssetTokenHelper';
 export * from './chain/tezos/contracts/tzip12/FA2Helper';
-export * from './chain/tezos/contracts/tzip12/FA2Types';
 export * from './chain/tezos/contracts/TzbtcTokenHelper';
 export * from './chain/tezos/contracts/WrappedTezosHelper';
